requirements.txt
pytest
numpy
python-dotenv
rich
. # self install for scripts